Add wi and wt commands; add hello-world program
Added the wi and wt commands.  wi is the interpreter and wt is a
translator.

The translator will translate between an "assembly" version of the
language and the vanilla whitespace code.

Both commands can take input from STDIN or a file and write output to
STDOUT or a file.  Default is to use STDIN/STDOUT if no file specified.

The hello-world program is code copied from the wikipedia page:
https://en.wikipedia.org/wiki/Whitespace_(programming_language)#Sample_code