Implement store/load; fix math
Implemented store and load.  Store takes an address and value off the
stack and puts it into a heap.  Load takes the address off the stack and
puts the value from the heap on the top of the stack.

Fixed a bunch of the math operations.  Turns out the operands were
backwards for most of them.

Fixed the implentation of readchar/readnumber.  These put values into
the heap, not on the stack.